# src/COP.API/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "0.1.0",
    "title" : "MONICA COP API"
  },
  "host" : "portal.monica-cloud.eu",
  "basePath" : "/cop/api",
  "schemes" : [ "https" ],
  "consumes" : [ "application/json" ],
  "produces" : [ "application/json" ],
  "paths" : {
...
    "/thing" : {
      "post" : {
        "tags" : [ "Thing" ],
        "description" : "add a new thing",
        "parameters" : [ {
          "in" : "body",
          "name" : "name",
          "description" : "name of thing",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Thing"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/GeneralResponse"
            }
          },
          "default" : {
            "description" : "Error",
            "schema" : {
              "$ref" : "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/things" : {
      "get" : {
        "tags" : [ "Thing" ],
        "description" : "lista all things",
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/GetThingListResponse"
            }
          },
          "default" : {
            "description" : "Error",
            "schema" : {
              "$ref" : "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/things/{id}" : {
      "get" : {
        "tags" : [ "Thing" ],
        "description" : "get thing by id",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/Thing"
            }
          },
          "default" : {
            "description" : "Error",
            "schema" : {
              "$ref" : "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Thing" ],
        "description" : "update Thing",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "updates",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Thing"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/GeneralResponse"
            }
          },
          "default" : {
            "description" : "Error",
            "schema" : {
              "$ref" : "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Thing" ],
        "description" : "delete thing by id",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/GeneralResponse"
            }
          },
          "default" : {
            "description" : "Error",
            "schema" : {
              "$ref" : "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "Thing" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "description" : "Thing id"
        },
        "name" : {
          "type" : "string",
          "description" : "Thing name"
        },
        "descriptiom" : {
          "type" : "string",
          "description" : "Thing description"
        },
        "thingtype" : {
          "type" : "integer",
          "description" : "Thing type"
        },
        "thingTemplate" : {
          "type" : "string",
          "description" : "Thing template"
        },
        "status" : {
          "type" : "integer",
          "description" : "Thing status"
        },
        "lat" : {
          "type" : "number",
          "description" : "Thing latitude"
        },
        "lont" : {
          "type" : "number",
          "description" : "Thing latitude"
        }
      },
      "example" : {
        "name" : "name",
        "thingtype" : 6,
        "id" : 0,
        "thingTemplate" : "thingTemplate",
        "lont" : 5.63737665663332876420099637471139430999755859375,
        "descriptiom" : "descriptiom",
        "lat" : 5.962133916683182377482808078639209270477294921875,
        "status" : 1
      }
    },
...
    "GetThingListResponse" : {
      "type" : "array",
      "items" : {
        "$ref" : "#/definitions/Thing"
      }
    },
...
    "GeneralResponse" : {
      "type" : "object",
      "required" : [ "description", "success" ],
      "properties" : {
        "success" : {
          "type" : "boolean",
          "description" : "returns 1 if successful"
        },
        "description" : {
          "type" : "string",
          "description" : "a short comment"
        }
      },
      "example" : {
        "success" : true,
        "description" : "description"
      }
    },
    "ErrorResponse" : {
      "required" : [ "message" ],
      "properties" : {
        "message" : {
          "type" : "string"
        }
      }
    }
  }
}